refactor: rewrite copy folder method to be more universal

+++ 04-copy-directory/index.js
@@ -1,25 +1,38 @@
 const fs = require('fs');
 const path = require('path');
 
-const copyDir = () => {
-  const folderPath = path.join(__dirname, 'files');
-  const copyFolderPath = path.join(__dirname, 'files-copy');
+const source = path.join(__dirname);
+const destination = path.join(__dirname);
 
-  fs.mkdir(copyFolderPath, (err) => {
-    if (err) throw err;
-  });
+copyDir(source, destination, 'files');
 
-  fs.promises.readdir(folderPath).then((files) => {
-    files.map((file) => {
-      fs.copyFile(
-        path.join(folderPath, file),
-        path.join(copyFolderPath, file),
-        (err) => {
-          if (err) throw err;
-        },
-      );
+async function copyDir(source, destination, folder) {
+  let folderName = folder;
+  if (source === destination) {
+    folderName += '-copy';
+  }
+  const destinationFolder = path.join(destination, folderName);
+  const sourceFolder = path.join(source, folder);
+
+  fs.rm(destinationFolder, { recursive: true, force: true }, () => {
+    fs.mkdir(destinationFolder, { recursive: true }, (err) => {
+      if (err) throw err;
+    });
+    fs.promises.readdir(sourceFolder).then((files) => {
+      files.map(async (file) => {
+        const stat = await fs.promises.lstat(path.join(sourceFolder, file));
+        if (stat.isFile()) {
+          fs.copyFile(
+            path.join(sourceFolder, file),
+            path.join(destinationFolder, file),
+            (err) => {
+              if (err) throw err;
+            },
+          );
+        } else {
+          copyDir(sourceFolder, destinationFolder, file);
+        }
+      });
     });
   });
-};
-
-copyDir();
+}
